asks user to resync file if outline was modified between keystrokes
If you make a modification to the file, leave the browser open, go to
another browser and modify the file there, then come BACK to the first
browser, it will ask if you want to resync the file since it was
modified externally.

This only occurs if the time between modifications is 1 hour. This will
eventually be made into a user preference.

References: #6