use last save time to determine outline resync
Instead of relying on the last keystroke time, which breaks if you are
using some kind of tab suspension, we keep track of the last save time.
This way, if the last save time is over 1 hour, we can ask if the user
wants to resync based on the last file.