bug: deleting first child node with siblings loses cursor
When we delete the first child node that has siblings, when we attempt
to get the `prevSibling` to place the cursor on - it technically selects
a `nodeContent` container that doesn't have a `data-id` attribute.

This fix ensures that it is only possible to select nodes that have a
`data-id` attribute.