bug: only account for if the top of an elemnet is out of view
We were originally checking if the whole element was out of view, but
due to the nested nature of things, this was always. Instead we only
check to see if the top of the element is out of view. If it is, we
scroll it into view otherwise we assume that enough of the node is
visible for you.