display/retrieve outline from server
After logging in you get prompted to load a remote outline or continue
with the one you were using before. If you select the remote outline, it
will overwrite your local copy with the remote one.

If you select cancel, it will NOT load the remote outline, but instead
attempt to sync your local outline remotely.