add sidebar + initial system for decoupling keyboard events
Right now all keyboardjs events are listed in a giant file and it makes
navigation a bit confusing. The new system allows us to define each
keyboard even indepedently (which incidentally makes testing easier),
and then dynamically maps all of them in the client.

This can then be further imported to generate the help docs.