initial pass at blog publishing
This required us to display a permanent "node details" page that
displays things like the ID/archive state. Eventually it will also
display whether or not the node is published, and `$metadata` that is
related to the post so that you can get an overview from one section.

The next step was a hard-coded markdown export. Right now it's listed as
a "Hugo Integration".. but really it's a markdown exporter and the only
"hugo" specific thing about it is the frontmatter generation.

It will remain in this branch for some time until I can iron more bugs
out