clean up and bug hunting on image publishing
You can now drag an image into the outliner and it will copy it to
$APPLOCALDATA/images, time-stamping to prevent duplicates. The image
will display in the node below the currently selected node.

when you publish, images are copied to the hugo image directory defined
in the configuration.

By default the permissions are on /home/ -> but this will need to be
tweaked further.