feat: Themes (#22)
* feat: shift+enter while editing adds a new line to the node

instead of creating a node, shift+enter adds a new line, allowing for
multi-line nodes.

* feat: themes and their manager!

this adds a simple theme manager (ctrl+t) and two themees (default and
dark) that the user can select between. It will remmeber your choice
between restarts of the application and apply across all your outlines.

AI Usage:
This update is a weird one. We utilized kiro to build the design plan
and then claude code to generate the bulk of the functionality, before
hand-editing.

I like the process of kiro to generate a more detailed plan that I can
go in and tweak before handing it off to claude - it feels more "full
featured" than just asking claude to go crazy.

* chore: update tauri